Force STOMP Reconnections on Selector Change
There is a bug, suspected to be related to backend buffering in
the STOMP 1.0 protocol, that causes selectors to not correctly
pick up messages after the first time. E.g. changing a selector
on an open connection will not show new messages.

STOMP 1.1+ clients should not need this because they do per
subscription tracking of messages.

Change-Id: Ib57f8650551f6a5090ef60742ed6ef22275be641